Introduced "regions" to bzfs.  A region is a volume of space.  This change uses regions to define the location of team bases, obstacles, and where flags can be inserted (i.e. spawned).
Regions are either a single shape (obstacles must be single shapes)
or an expression of shapes using union, difference, and intersection.
The shapes can be any type derived from Shape, currently boxes and
pyramids.

This change splits world creation and reading out of bzfs.cxx into
separate files.  It also adds code to generate XML for regions from
the world then parses that XML to create the regions.  The regions
are then used to insert/drop flags.  This is in anticipation of
changing world files to XML.

Bases will soon no longer be obstacles but just regions where flags
can be captured.  That means floating bases will need obstacles
below them for players to drive on.  That also requires support for
flags on obstacles or team flags would fall to the ground.  So this
change also discards the old flagsOnBuildings code (which didn't
work right anyway) and fully supports flags on buildings.  It
doesn't allow flags on pyramids because it's extremely hard to grab
flags on pyramids.

git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@561 08b3d480-bf2c-0410-a26f-811ee3361c24