Only players with POLL permission are eligible to take part in a vote
A cheat is going around which loads a bunch of fake players on the
server - if these are valid accounts (but not identified) then they
artificially increase the number of available players for the poll so
that it is impossible to make the poll succeed.  This effectively
makes player initiated polls useless.

Now we only consider verified players with poll permission as
available to vote - the extra observers/bots the cheater loads does
not affect the poll outcome.

git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@21423 08b3d480-bf2c-0410-a26f-811ee3361c24