A way to fire a world weapon and get the shot ID
A new version of fire world weapon that takes a pointer to a shot ID and
gives back the first dead shot ID based on the lifetime from all other
shots ( uses shot manager). Should make it easier for people to have
multiple plug-ins that want to use world weapons and not have them
conflict in IDs. There really isn't any reason for a plug-in to set the
shot ID, it can always keep a map to associate it with a player.