Fix mouse wheel firing causing MG to endlessly fire (#316)
Our mouse wheel event only triggered a KeyDown event. When mouse wheel
was mapped to fire, this would cause MG to fire endlessly. This change
adds a custom user event that gets pushed to SDL when a mouse wheel
event is triggered. When this custom event is processed, we trigger a
KeyUp event for the same mouse wheel direction.