Disable resizable SDL1 windowed mode
As pointed to in issue #303, some fixes we made for SDL2 broke the
ability to resize the window when building against SDL1. The issue seems
like the OpenGL viewport isn't adjusting to the window size correctly.
We no longer use SDL1 by default on any platform and we don't really
want to dedicate any further time to it. So I'm just disabling the
ability to resize the window when built against SDL1. It's still
possible to launch windowed at a specific size using the -window
argument.