Update CodeQL workflow (#317)
The main purpose of this change is to switch to using the v2 CodeQL actions since the v1 version is deprecated. This also adjusts the build process to build the entire codebase instead of just bzfs and a basic bzadmin. This now reflects our normal Linux build process including the client with SDL2 and bzadmin with ncurses. It will also try to use multiple CPU cores if available.

Additionally, this add the master branch to the list of push and pull-request branches that CodeQL should analyze, adds the write permission for security events, switches to the v3 checkout action, and removes the depth and checkout settings as recommended.