Docker: Multi-stage build, default version, custom plugins, data volume
The Dockerfile now uses a multi-stage build which compiles the source in
one image and then copies binaries into a second clean image. This
reduces the final image size from 758MB to 67.5MB, and is further
reduced to 19.2MB by striping the compiled binaries.

The branch/tag of the Git checkout was required to be specified
previously, but now has a default of 2.4 if not provided.

Custom plugins are automatically copied into the build image and
compiled by including them in a custom_plugins directory.

A directory/volume is created at /data in the final image which can be
mapped to a directory on the host system for configuration, world,
groups, and other files.