Simplify the Dockerfile build
Instead of using git inside the build container, just copy the source
directory from the host system. This makes it much easier to test local
changes since it does not need to be committed and pushed first. It also
makes the build process more familiar since custom plugins can just be
added to the plugins/ directory like normal, and configure flags can be
passed to the 'configure' build argument.