Second attempt at fixing skewed textures.
Fixes issue #346, which reported that the skewed textures occured if the
game was toggled between fullscreen and windowed, or if the resolution
changed. The GL_PACK_ALIGNMENT/GL_UNPACK_ALIGNMENT values needed to be
set before ContextInitializer::executeInitFuncs(), or else affected
textures would be skewed upon context re-creation. The moves the calls
to glPixelStorei setting the alignment into the
OpenGLGState::initGLState() function.